Add some changes

+++ src/pages/reservations/Components/Reservation.tsx
@@ -5,12 +5,12 @@ import {
     deleteReservationById,
     getReservation,
 } from '../../../store/slices/Reservations';
-
 import trash from '../../../assets/images/trash.png';
 import update from '../../../assets/images/edit.png';
 import view from '../../../assets/images/find.png';
 import { useNavigate } from 'react-router-dom';
 import { updateOwnerCar } from '../../../store/slices/Cars';
+import Swal from 'sweetalert2';
 
 interface Props {
     reservation: ReservationModel;
@@ -26,9 +26,21 @@ export const Reservation: React.FC<Props> = ({ reservation }) => {
     }
 
     function handleRemove(idRental: number, idCar: number) {
-        // console.log('go-remove');
-        dispatch(deleteReservationById(idRental));
-        dispatch(updateOwnerCar({ idCar }));
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!',
+        }).then((result) => {
+            if (result.isConfirmed) {
+                dispatch(deleteReservationById(idRental));
+                dispatch(updateOwnerCar({ idCar }));
+                Swal.fire('Saved!', '', 'success');
+            }
+        });
     }
 
     return (
